mapping + gui update

+++ Xi0nGUI/src/robot_directing/decisional/StateMachineTransitionForDecisionV2.java
@@ -34,7 +34,7 @@ public class StateMachineTransitionForDecisionV2 {
     // SENSORS ----------------------------
     // ------------------------------------
 	
-	private int rightSideDistance;
+	private float rightSideDistance;
 	private float frontalDistance;
 	private float servoAngle;
