+++ Xi0nGUI/src/robot_directing/decisional/StateMachineTransitionForDecisionV4.java
@@ -37,7 +37,7 @@ public class StateMachineTransitionForDecisionV4 {
     // SENSORS ----------------------------
     // ------------------------------------
 	
-	private int rightSideDistance;
+	private float rightSideDistance;
 	private float frontalDistance;
 	private float servoAngle;
