+++ Xi0nGUI/src/robot_directing/decisional/StateMachineTransitionForDecisionV5.java
@@ -39,7 +39,7 @@ public class StateMachineTransitionForDecisionV5 {
     // SENSORS ----------------------------
     // ------------------------------------
 	
-	private int rightSideDistance;
+	private float rightSideDistance;
 	private float frontalDistance;
 	private float servoAngle;
